move CombineObjects function to a public class
