+++ com.unity.probuilder/Runtime/MeshOperations/CombineMeshes.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bb36fee1f817e4860a9937f064b1a192
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
